Implement support for playing a card
This method also errors out if this exact card has already
been played. It also errors out if the game is finished, i.e.
all of the cards have been played.

It creates the new hand as well because the game state
assumes that a new hand will always exist.