Clean up how defining of the briscola suit works
It will require explicit setting of a partner card after bidding
is complete.

There is also the TODO of defining the dealer and maintaining
playing order. Going to with each new person to join is to the
right of the previous person.

Unfortunately, I have introduced a few N + 1 queries, especially
in the winner and score computation.

I'm also missing transactions. I want to first get the basic
gameplay working, then build out a basic UI, and then I can
revisit the poor query performance. It is worth noting that this
is not meant to be a performant app so I can at least start
using it and then rewrite it while it's in use.