Implement the ability to recreate a new game with the same players
This is mostly for convenience so that you don't have to go to
the create game screen again.

This commit also fixes a bug that occurred when you played the
last card in a hand and won the hand. In that case, your previous
card will still be stored as the selected card and you would be
asked if you wanted to play it. This commit clears the selected
card after playing.